+++ src/bin/devmgr/src/nvme_test.rs
@@ -4,7 +4,7 @@ use twizzler_driver::bus::pcie::PcieDeviceInfo;
 struct NvmeController {}
 
 pub fn start() {
-    let device_root = twizzler_driver::device::get_bustree_root();
+    let device_root = twizzler_driver::get_bustree_root();
     for device in device_root.children() {
         if device.is_bus() && device.bus_type() == BusType::Pcie {
             for child in device.children() {
